i18Setup
Added i18n

Installed i18next as the i18n provider and its dependencies.

i18n Middleware

i18n Setup
Created a layout to provide the i18n to the whole app.
This layout only works for Client Components.
Pending to expand to server components.

Page.tsx and the messages->*.json are there for the options but contain
only an example.